+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ CHANGELOG
 5.3.57 (unreleased)
 -------------------
 
+- Add metric for object cache records size [lferran]
+
 - Fix metrics context manager when prometheus not installed [lferran]
 
 5.3.56 (2020-10-14)